We got noise sheeple! We got em!

+++ app/src/main/cpp/fire/simulation/wavelet_turbulence.h
@@ -33,7 +33,7 @@ class WaveletTurbulence {
     DataTexturePair* energy;
     DataTexturePair* texture_coord;
 
-    unsigned int band_min, band_max;
+    float band_min, band_max;
 
 public:
     int init(SlabOperator* slab);
